refactor: divide game details into separate containers
